Y flipping and synchronisation
Preparing for the next release features that are :

* Fixing a bug where Quest users saw the image
  flipped on the Y-Axis. The shader used now handles
  the flipping correctly.

* The slideshow handles synchronisation if desired.

That should be it for this release.
The next one will add the Play/Stop buttons so that
people can stop the slideshow.

I still don't know how I'm going to handle this gracefully,
given that we're using a custom error check loop here,
thanks to the VRChat Geniuses who made it so that
the basic error handling is called from a thread that
crashes Unity if you touch anything Unity related.

So I'd like to be able to stop the Error check loop,
while avoid executing the error check loop twice.

Signed-off-by: Voyage <voyage@miouyouyou.fr>