Simple Diaporama : 1.0 beta 1
The name of the DLL is still Simple Slideshow,
but let's be clear, tihs is a slideshow.

I'll try to release it before the end of the week.

This has been tested for more than 1 month in my world,
and all the issues have been resolved.

Notably :
* Downloaders now are recycled very early, in order to
avoid GPU memory leaks.
  For some reasons, if you continue to use the same
  ImageDownloader, all the downloaded pictures will be
  kept in GPU memory, killing Quest users very early.

* Double buffering downloaders, in order to avoid a black
  square everytime a downloader gets recycled when
  proceeding to the next download.

* Quest users see the images correctly.

Signed-off-by: Voyage <voyage@miouyouyou.fr>